+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.95
+### 🚀 New components 🚀
+#### **coralogix-aws-shipper**
+- Add new variables custom_csv_header and custom_metadata
+
 ## v1.0.94
 #### **coralogix-aws-shipper**
 ### 🧰 Bug fixes 🧰
@@ -31,7 +36,7 @@
 
 ## v1.0.90
 ### 💡 Enhancements 💡
-#### **coralogix-aws-shipper**
+
 - [cds-1050] add support for x86 to template
 
 ## v1.0.89